+++ src/main/java/com/music/kevinmusic/service/UserService.java
@@ -3,7 +3,7 @@ package com.music.kevinmusic.service;
 import com.music.kevinmusic.command.PointCommand;
 import com.music.kevinmusic.command.UserCommand;
 import com.music.kevinmusic.domain.Information;
-import com.music.kevinmusic.domain.User;
+import com.music.kevinmusic.domain.security.User;
 import com.music.kevinmusic.request.UserRequest;
 import org.springframework.data.domain.Page;
 
@@ -13,7 +13,7 @@ public interface UserService {
 
     void saveAll(List<User> list);
 
-    User getUserById(Long id);
+    User getUserById(Integer id);
 
     User getUserByUsername(String username);
